+++ metadata.rb
@@ -7,4 +7,5 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.5.1'
 
 depends 'java'
+depends 'ulimit'
 recommends 'authbind'