+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.5.2:
 
+* Ensure that the ulimits are correctly set for elasticsearch user.
 * Remove non US-ASCII character in elasticsearch.yml
 * Correct the configuration so that it is possible to change the bind interface correctly.
 * Upgrade to Elasticsearch 0.20.6.